refactor(client): split error.rs so it stays under the 500-line limit
CI's Format job failed on PR #105: `crates/a2a-protocol-client/src/error.rs`
had reached 559 lines and is not in `.file-length-baseline`. The four
stream-lag tests added in 6a775dc tipped a file that was already sitting at
~499 over the limit.

Split rather than baselined, which is what CONTRIBUTING asks for when the
split is natural: production code stays in `error/mod.rs` (195 lines), the
`#[cfg(test)]` module becomes `error/tests.rs` (373). Nothing about the public
surface changes — no item moved, only the test module.

I had run this check before opening the PR and misread it. The command was
`./scripts/check_file_lengths.sh 2>&1 | tail -1`, which printed "makes it a
visible decision instead of a silent one." — a line from the script's
*failure* message — and I took it for output rather than a verdict, because I
piped away everything else and never looked at the exit code. Truncating a
gate's output and not checking its status is the same defect this branch
spent its time removing from the harnesses.

All four Format-job steps now verified individually by exit code:
cargo fmt --check 0, check_proto_copies.sh 0, check_file_lengths.sh 0,
prove_workflow_gates_fail.py 0. Also clippy 0, cargo doc 0, and
cargo test --workspace --all-features 2699 passed / 0 failed.
